+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 * Implementation of unmanaged applications.
 * Added two binary parts message for publisher/subscriber and requester/responder.
 * Refined internal requester port name by adding the responder id.
+* Added connectToRequester method to Request.